cleanup

+++ modules/deploy/data/03-letsencrypt.sh
@@ -11,23 +11,23 @@ source /provision/vars.sh
 
 if ! is_controller; then
     echo "not a controller, exiting"
-    return
+    exit
 fi
 
-# define these in extras!
+# Define these in extras!
 # CFG_UN=""
 # CFG_PW=""
 # CFG_HN=""
 # CFG_EMAIL=""
 
-# if there's no hostname then return immediately (these scripts are sourced)
+# If there's no hostname then exit immediately
 if [ -z "${CFG_HN}" ]; then
     echo "no hostname configured, exiting"
-    return
+    exit
 fi
 
-# update our hostname on DynDNS
-IP=$(curl -s4 canhazip.com)
+# Update our hostname on DynDNS
+IP=$(curl -s4 canhazip.com) || exit
 auth=$(echo -n "$CFG_UN:$CFG_PW" | base64)
 attempts=5
 while [ $attempts -gt 0 ]; do
@@ -44,12 +44,10 @@ while [ $attempts -gt 0 ]; do
     ((attempts--))
 done
 
-echo
-
 copyfile ${CFG_HN}-fullchain.pem /tmp/fullchain.pem
 copyfile ${CFG_HN}-privkey.pem /tmp/privkey.pem
 
-if openssl x509 -text </tmp/fullchain.pem | grep ${CFG_HN}; then
+if test -f /tmp/fullchain.pem && openssl x509 -text </tmp/fullchain.pem | grep -q ${CFG_HN}; then
     mkdir -p /etc/letsencrypt/live/$CFG_HN
     mv /tmp/fullchain.pem /etc/letsencrypt/live/$CFG_HN/fullchain.pem
     mv /tmp/privkey.pem /etc/letsencrypt/live/$CFG_HN/privkey.pem
@@ -67,33 +65,21 @@ ln -s /snap/bin/certbot /usr/bin/certbot
 
 # ONLY request a cert if there's not already one present!
 if ! [ -d /etc/letsencrypt/live/$CFG_HN ]; then
-    echo "### before"
-    /usr/bin/firewall-cmd --list-all
-    echo "### apply"
     /usr/bin/firewall-cmd --reload
-    echo "### after"
-    /usr/bin/firewall-cmd --list-all
-    echo "### done"
     /usr/bin/certbot --domain $CFG_HN --noninteractive --nginx certonly
-    # could copy back to the cloud storage here? According to the naming
-    # scheme hostname-fullchain.pem and hostname-privkey.pem
 fi
 
-# copy the cert to the nginx configuration
+# Copy the cert to the nginx configuration
 mkdir -p /etc/nginx/oldcerts
 mv /etc/nginx/*.pem /etc/nginx/oldcerts/
 cp /etc/letsencrypt/live/$CFG_HN/fullchain.pem /etc/nginx/pubkey.pem
 cp /etc/letsencrypt/live/$CFG_HN/privkey.pem /etc/nginx/privkey.pem
 
-# Cockpit loads certificates from the /etc/cockpit/ws-certs.d directory. It
-# will use the last file in the directory with a .cert or .crt extension in
-# alphabetical order. The private key can be contained in a separate file with
-# the same name as the certificate, but with a .key suffix instead. The key
-# must not be encrypted.
+# Cockpit
 rm /etc/cockpit/ws-certs.d/*
 cp /etc/letsencrypt/live/$CFG_HN/fullchain.pem /etc/cockpit/ws-certs.d/$CFG_HN.crt
 cp /etc/letsencrypt/live/$CFG_HN/privkey.pem /etc/cockpit/ws-certs.d/$CFG_HN.key
 
-# reload affected services
+# Reload affected services
 systemctl reload nginx
 systemctl restart cockpit
